Atualização filtro variação do ativo

+++ SS.Tecnologia.YahooFinance/Models/Quote.cs
@@ -11,7 +11,7 @@ namespace SS.Tecnologia.YahooFinance.Models
         public List<double?>? Open { get; set; }
         public List<double?>? Low { get; set; }
         public List<double?>? High { get; set; }
-        public List<int?>? Volume { get; set; }
+        public List<long?>? Volume { get; set; }
         public List<double?>? Close { get; set; }
     }
 }
